+++ src/components/Publisher.vue
@@ -2,13 +2,14 @@
   <div class="mqtt-client__publisher col-md-6 col-sm-12 col-xs-12">
     <q-card class="publisher__item q-ma-sm">
       <q-card-actions>
-        <q-item style="width: calc(100% - 20px);" dense class="q-card cursor-pointer" highlight @click.native="$emit('publish')">
-          <q-item-main>
-            <q-item-tile style="font-size: 0.9rem;" class="uppercase text-bold" label>Publish message</q-item-tile>
-            <q-item-tile style="margin-top: 0; font-size: 0.75rem" class="ellipsis" sublabel>{{config.topic}}</q-item-tile>
-          </q-item-main>
-          <q-item-side right icon="mdi-send"/>
-        </q-item>
+        <q-btn :disable="!config.topic" align="left" style="width: calc(100% - 20px);" @click="$emit('publish')" icon-right="mdi-send">
+          <q-item style="width: calc(100% - 32px);" dense class="q-px-none">
+            <q-item-main>
+              <q-item-tile style="font-size: 0.9rem;" class="uppercase text-bold" label>Publish message</q-item-tile>
+              <q-item-tile style="margin-top: 0; font-size: 0.75rem; text-transform: initial;" class="ellipsis" sublabel>{{config.topic}}</q-item-tile>
+            </q-item-main>
+          </q-item>
+        </q-btn>
       </q-card-actions>
       <q-icon class="publisher__remove cursor-pointer" size="1rem" name="mdi-close" @click.native="$emit('remove')"/>
       <q-card-main class="publisher__main q-pb-none">
@@ -25,7 +26,7 @@
             <q-checkbox style="display: flex;" color="dark" class="q-mt-sm q-mb-sm" v-model="config.options.retain" label="Retain"/>
             <q-checkbox style="display: flex;" color="dark" class="q-mt-sm q-mb-sm" v-model="config.options.dup" label="Duplicate flag"/>
             <q-collapsible v-if="version === 5" class="q-mt-sm q-mb-sm bg-grey-4" label="Properties">
-              <q-checkbox color="dark" class="q-mt-sm q-mb-sm" v-model="config.options.properties.payloadFormatIndicator" label="Payload format indicator"/>
+              <q-checkbox color="dark" class="q-mt-sm q-mb-sm" toggle-indeterminate :indeterminate-value="null" v-model="config.options.properties.payloadFormatIndicator" label="Payload format indicator"/>
               <q-input color="dark" type="number" v-model="config.options.properties.messageExpiryInterval" float-label="Message expiry interval"/>
               <q-input color="dark" type="number" v-model="config.options.properties.topicAlias" float-label="Topic alias"/>
               <q-input color="dark" v-model="config.options.properties.responseTopic" float-label="Response topic"/>
@@ -44,7 +45,6 @@
                   <q-btn :disable="!publishUserProperty.name || !publishUserProperty.value" style="width: 100%" class="q-mt-sm" color="dark" @click="addPublishUserProperty()">Add</q-btn>
                 </div>
               </div>
-              <q-input color="dark" type="number" v-model="config.options.properties.subscriptionIdentifier" float-label="Subscription identifier"/>
               <q-input color="dark" v-model="config.options.properties.contentType" float-label="Content type"/>
             </q-collapsible>
           </q-collapsible>
@@ -96,6 +96,12 @@ export default {
       handler (val) {
         this.$emit('input', val)
       }
+    },
+    value: {
+      deep: true,
+      handler (value) {
+        this.config = value
+      }
     }
   }
 }
@@ -103,7 +109,6 @@ export default {
 
 <style lang="stylus">
   .mqtt-client__publisher
-    transition all ease-in-out .3s
     .publisher__item
       border 2px solid indigo
       height calc(100% - 32px)